Move FinTsOptions and Credentials to Options namespace
Since library users don't need to interact with them anymore, it's better to make them less discoverable.
