+++ SamplesNew/login.php
@@ -11,7 +11,7 @@ require '../vendor/autoload.php';
 // The configuration options up here are considered static wrt. the library's internal state and its requests.
 // That is, even if you persist the FinTs instance, you need to be able to reproduce all this information from some
 // application-specific storage (e.g. your database) in order to use the phpFinTS library.
-$options = new \Fhp\FinTsOptions();
+$options = new \Fhp\Options\FinTsOptions();
 $url = ''; // HBCI / FinTS Url can be found here: https://www.hbci-zka.de/institute/institut_auswahl.htm (use the PIN/TAN URL)
 $bankCode = ''; // Your bank code / Bankleitzahl
 $productName = ''; // The number you receive after registration / FinTS-Registrierungsnummer
